Add support for async contexts
In async calls, since the mock instance being used is a copy of the
original instance, the state is not preserved outside the async context.
This commit addresses the issue. We store a map of all Umock instances
in a static map, keyed by a guid, and copy the current instance to the map
after each method call. The subsequent verification methods uses the instance
stored in the static map instead of the current instance.

Change all occurences of `isTest` to `IsTest` to conform with Java style guides
for annotations. Update Readme to note async behavior support.